backend/requirements.txt
Flask==3.1.2
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.44
cryptography==41.0.7
python-dotenv==1.0.0
Werkzeug==3.1.3
mysql-connector-python==9.5.0
transformers==5.1.0
datasets==4.6.0